+++ packages/remix-server-runtime/server.ts
@@ -18,19 +18,31 @@ import type { HandleErrorFunction, ServerBuild } from "./build";
 import type { EntryContext } from "./entry";
 import { createEntryRouteModules } from "./entry";
 import { sanitizeErrors, serializeError, serializeErrors } from "./errors";
-import { getDocumentHeadersRR } from "./headers";
+import { getDocumentHeaders } from "./headers";
 import invariant from "./invariant";
 import { ServerMode, isServerMode } from "./mode";
+import type { RouteMatch } from "./routeMatching";
 import { matchServerRoutes } from "./routeMatching";
-import type { ServerRoute } from "./routes";
+import type { EntryRoute, ServerRoute } from "./routes";
 import { createStaticHandlerDataRoutes, createRoutes } from "./routes";
 import {
   createDeferredReadableStream,
   isRedirectResponse,
   isResponse,
+  json,
 } from "./responses";
 import { createServerHandoffString } from "./serverHandoff";
 import { getDevServerHooks } from "./dev";
+import type { SingleFetchResult, SingleFetchResults } from "./single-fetch";
+import {
+  encodeViaTurboStream,
+  getSingleFetchRedirect,
+  singleFetchAction,
+  singleFetchLoaders,
+  SingleFetchRedirectSymbol,
+  SINGLE_FETCH_REDIRECT_STATUS,
+} from "./single-fetch";
+import { resourceRouteJsonWarning } from "./deprecations";
 
 export type RequestHandler = (
   request: Request,
@@ -41,6 +53,17 @@ export type CreateRequestHandlerFunction = (
   build: ServerBuild | (() => ServerBuild | Promise<ServerBuild>),
   mode?: string
 ) => RequestHandler;
+
+// Do not include a response body if the status code is one of these,
+// otherwise `undici` will throw an error when constructing the Response:
+//   https://github.com/nodejs/undici/blob/bd98a6303e45d5e0d44192a93731b1defdb415f3/lib/web/fetch/response.js#L522-L528
+//
+// Specs:
+//   https://datatracker.ietf.org/doc/html/rfc9110#name-informational-1xx
+//   https://datatracker.ietf.org/doc/html/rfc9110#name-204-no-content
+//   https://datatracker.ietf.org/doc/html/rfc9110#name-205-reset-content
+//   https://datatracker.ietf.org/doc/html/rfc9110#name-304-not-modified
+const NO_BODY_STATUS_CODES = new Set([100, 101, 204, 205, 304]);
 
 function derive(build: ServerBuild, mode?: string) {
   let routes = createRoutes(build.routes);
@@ -101,8 +124,7 @@ export const createRequestHandler: CreateRequestHandlerFunction = (
     }
 
     let url = new URL(request.url);
-
-    let matches = matchServerRoutes(routes, url.pathname, _build.basename);
+    let params: RouteMatch<ServerRoute>["params"] = {};
     let handleError = (error: unknown) => {
       if (mode === ServerMode.Development) {
         getDevServerHooks()?.processRequestError?.(error);
@@ -110,16 +132,45 @@ export const createRequestHandler: CreateRequestHandlerFunction = (
 
       errorHandler(error, {
         context: loadContext,
-        params: matches && matches.length > 0 ? matches[0].params : {},
+        params,
         request,
       });
     };
 
+    // Manifest request for fog of war
+
+    let manifestUrl = `${_build.basename ?? "/"}/__manifest`.replace(
+      /\/+/g,
+      "/"
+    );
+    if (url.pathname === manifestUrl) {
+      try {
+        let res = await handleManifestRequest(_build, routes, url);
+        return res;
+      } catch (e) {
+        handleError(e);
+        return new Response("Unknown Server Error", { status: 500 });
+      }
+    }
+
+    let matches = matchServerRoutes(routes, url.pathname, _build.basename);
+    if (matches && matches.length > 0) {
+      Object.assign(params, matches[0].params);
+    }
+
     let response: Response;
     if (url.searchParams.has("_data")) {
+      if (_build.future.v3_singleFetch) {
+        handleError(
+          new Error(
+            "Warning: Single fetch-enabled apps should not be making ?_data requests, " +
+              "this is likely to break in the future"
+          )
+        );
+      }
       let routeId = url.searchParams.get("_data")!;
 
-      response = await handleDataRequestRR(
+      response = await handleDataRequest(
         serverMode,
         _build,
         staticHandler,
@@ -132,17 +183,81 @@ export const createRequestHandler: CreateRequestHandlerFunction = (
       if (_build.entry.module.handleDataRequest) {
         response = await _build.entry.module.handleDataRequest(response, {
           context: loadContext,
-          params: matches?.find((m) => m.route.id == routeId)?.params || {},
+          params,
           request,
         });
+
+        if (isRedirectResponse(response)) {
+          response = createRemixRedirectResponse(response, _build.basename);
+        }
+      }
+    } else if (_build.future.v3_singleFetch && url.pathname.endsWith(".data")) {
+      let handlerUrl = new URL(request.url);
+      handlerUrl.pathname = handlerUrl.pathname
+        .replace(/\.data$/, "")
+        .replace(/^\/_root$/, "/");
+
+      let singleFetchMatches = matchServerRoutes(
+        routes,
+        handlerUrl.pathname,
+        _build.basename
+      );
+
+      response = await handleSingleFetchRequest(
+        serverMode,
+        _build,
+        staticHandler,
+        request,
+        handlerUrl,
+        loadContext,
+        handleError
+      );
+
+      if (_build.entry.module.handleDataRequest) {
+        response = await _build.entry.module.handleDataRequest(response, {
+          context: loadContext,
+          params: singleFetchMatches ? singleFetchMatches[0].params : {},
+          request,
+        });
+
+        if (isRedirectResponse(response)) {
+          let result: SingleFetchResult | SingleFetchResults =
+            getSingleFetchRedirect(
+              response.status,
+              response.headers,
+              _build.basename
+            );
+
+          if (request.method === "GET") {
+            result = {
+              [SingleFetchRedirectSymbol]: result,
+            };
+          }
+          let headers = new Headers(response.headers);
+          headers.set("Content-Type", "text/x-script");
+
+          return new Response(
+            encodeViaTurboStream(
+              result,
+              request.signal,
+              _build.entry.module.streamTimeout,
+              serverMode
+            ),
+            {
+              status: SINGLE_FETCH_REDIRECT_STATUS,
+              headers,
+            }
+          );
+        }
       }
     } else if (
       matches &&
       matches[matches.length - 1].route.module.default == null &&
       matches[matches.length - 1].route.module.ErrorBoundary == null
     ) {
-      response = await handleResourceRequestRR(
+      response = await handleResourceRequest(
         serverMode,
+        _build,
         staticHandler,
         matches.slice(-1)[0].route.id,
         request,
@@ -155,7 +270,7 @@ export const createRequestHandler: CreateRequestHandlerFunction = (
           ? await getDevServerHooks()?.getCriticalCss?.(_build, url.pathname)
           : undefined;
 
-      response = await handleDocumentRequestRR(
+      response = await handleDocumentRequest(
         serverMode,
         _build,
         staticHandler,
@@ -178,7 +293,44 @@ export const createRequestHandler: CreateRequestHandlerFunction = (
   };
 };
 
-async function handleDataRequestRR(
+async function handleManifestRequest(
+  build: ServerBuild,
+  routes: ServerRoute[],
+  url: URL
+) {
+  if (build.assets.version !== url.searchParams.get("version")) {
+    return new Response(null, {
+      status: 204,
+      headers: {
+        "X-Remix-Reload-Document": "true",
+      },
+    });
+  }
+
+  let patches: Record<string, EntryRoute> = {};
+
+  if (url.searchParams.has("p")) {
+    for (let path of url.searchParams.getAll("p")) {
+      let matches = matchServerRoutes(routes, path, build.basename);
+      if (matches) {
+        for (let match of matches) {
+          let routeId = match.route.id;
+          patches[routeId] = build.assets.routes[routeId];
+        }
+      }
+    }
+
+    return json(patches, {
+      headers: {
+        "Cache-Control": "public, max-age=31536000, immutable",
+      },
+    }) as Response; // Override the TypedResponse stuff from json()
+  }
+
+  return new Response("Invalid Request", { status: 400 });
+}
+
+async function handleDataRequest(
   serverMode: ServerMode,
   build: ServerBuild,
   staticHandler: StaticHandler,
@@ -194,27 +346,7 @@ async function handleDataRequestRR(
     });
 
     if (isRedirectResponse(response)) {
-      // We don't have any way to prevent a fetch request from following
-      // redirects. So we use the `X-Remix-Redirect` header to indicate the
-      // next URL, and then "follow" the redirect manually on the client.
-      let headers = new Headers(response.headers);
-      let redirectUrl = headers.get("Location")!;
-      headers.set(
-        "X-Remix-Redirect",
-        build.basename
-          ? stripBasename(redirectUrl, build.basename) || redirectUrl
-          : redirectUrl
-      );
-      headers.set("X-Remix-Status", response.status);
-      headers.delete("Location");
-      if (response.headers.get("Set-Cookie") !== null) {
-        headers.set("X-Remix-Revalidate", "yes");
-      }
-
-      return new Response(null, {
-        status: 204,
-        headers,
-      });
+      return createRemixRedirectResponse(response, build.basename);
     }
 
     if (DEFERRED_SYMBOL in response) {
@@ -236,18 +368,16 @@ async function handleDataRequestRR(
 
     // Mark all successful responses with a header so we can identify in-flight
     // network errors that are missing this header
-    response.headers.set("X-Remix-Response", "yes");
+    response = safelySetHeader(response, "X-Remix-Response", "yes");
     return response;
   } catch (error: unknown) {
     if (isResponse(error)) {
-      error.headers.set("X-Remix-Catch", "yes");
-      return error;
+      let response = safelySetHeader(error, "X-Remix-Catch", "yes");
+      return response;
     }
 
     if (isRouteErrorResponse(error)) {
-      if (error) {
-        handleError(error);
-      }
+      handleError(error);
       return errorResponseToJson(error, serverMode);
     }
 
@@ -265,7 +395,69 @@ async function handleDataRequestRR(
   }
 }
 
-async function handleDocumentRequestRR(
+async function handleSingleFetchRequest(
+  serverMode: ServerMode,
+  build: ServerBuild,
+  staticHandler: StaticHandler,
+  request: Request,
+  handlerUrl: URL,
+  loadContext: AppLoadContext,
+  handleError: (err: unknown) => void
+): Promise<Response> {
+  let { result, headers, status } =
+    request.method !== "GET"
+      ? await singleFetchAction(
+          build,
+          serverMode,
+          staticHandler,
+          request,
+          handlerUrl,
+          loadContext,
+          handleError
+        )
+      : await singleFetchLoaders(
+          build,
+          serverMode,
+          staticHandler,
+          request,
+          handlerUrl,
+          loadContext,
+          handleError
+        );
+
+  // Mark all successful responses with a header so we can identify in-flight
+  // network errors that are missing this header
+  let resultHeaders = new Headers(headers);
+  resultHeaders.set("X-Remix-Response", "yes");
+
+  // Skip response body for unsupported status codes
+  if (NO_BODY_STATUS_CODES.has(status)) {
+    return new Response(null, { status, headers: resultHeaders });
+  }
+
+  // We use a less-descriptive `text/x-script` here instead of something like
+  // `text/x-turbo` to enable compression when deployed via Cloudflare.  See:
+  //  - https://github.com/remix-run/remix/issues/9884
+  //  - https://developers.cloudflare.com/speed/optimization/content/brotli/content-compression/
+  resultHeaders.set("Content-Type", "text/x-script");
+
+  // Note: Deferred data is already just Promises, so we don't have to mess
+  // `activeDeferreds` or anything :)
+  return new Response(
+    encodeViaTurboStream(
+      result,
+      request.signal,
+      build.entry.module.streamTimeout,
+      serverMode
+    ),
+    {
+      status: status || 200,
+      headers: resultHeaders,
+    }
+  );
+}
+
+async function handleDocumentRequest(
   serverMode: ServerMode,
   build: ServerBuild,
   staticHandler: StaticHandler,
@@ -288,10 +480,17 @@ async function handleDocumentRequestRR(
     return context;
   }
 
+  let headers = getDocumentHeaders(build, context);
+
+  // Skip response body for unsupported status codes
+  if (NO_BODY_STATUS_CODES.has(context.statusCode)) {
+    return new Response(null, { status: context.statusCode, headers });
+  }
+
   // Sanitize errors outside of development environments
   if (context.errors) {
     Object.values(context.errors).forEach((err) => {
-      // @ts-expect-error This is "private" from users but intended for internal use
+      // @ts-expect-error `err.error` is "private" from users but intended for internal use
       if (!isRouteErrorResponse(err) || err.error) {
         handleError(err);
       }
@@ -299,25 +498,37 @@ async function handleDocumentRequestRR(
     context.errors = sanitizeErrors(context.errors, serverMode);
   }
 
-  let headers = getDocumentHeadersRR(build, context);
-
+  // Server UI state to send to the client.
+  // - When single fetch is enabled, this is streamed down via `serverHandoffStream`
+  // - Otherwise it's stringified into `serverHandoffString`
+  let state = {
+    loaderData: context.loaderData,
+    actionData: context.actionData,
+    errors: serializeErrors(context.errors, serverMode),
+  };
   let entryContext: EntryContext = {
     manifest: build.assets,
     routeModules: createEntryRouteModules(build.routes),
     staticHandlerContext: context,
     criticalCss,
     serverHandoffString: createServerHandoffString({
-      url: context.location.pathname,
       basename: build.basename,
       criticalCss,
-      state: {
-        loaderData: context.loaderData,
-        actionData: context.actionData,
-        errors: serializeErrors(context.errors, serverMode),
-      },
       future: build.future,
       isSpaMode: build.isSpaMode,
+      ...(!build.future.v3_singleFetch ? { state } : null),
     }),
+    ...(build.future.v3_singleFetch
+      ? {
+          serverHandoffStream: encodeViaTurboStream(
+            state,
+            request.signal,
+            build.entry.module.streamTimeout,
+            serverMode
+          ),
+          renderMeta: {},
+        }
+      : null),
     future: build.future,
     isSpaMode: build.isSpaMode,
     serializeError: (err) => serializeError(err, serverMode),
@@ -340,21 +551,8 @@ async function handleDocumentRequestRR(
     // If they threw a response, unwrap it into an ErrorResponse like we would
     // have for a loader/action
     if (isResponse(error)) {
-      let data;
       try {
-        let contentType = error.headers.get("Content-Type");
-        // Check between word boundaries instead of startsWith() due to the last
-        // paragraph of https://httpwg.org/specs/rfc9110.html#field.content-type
-        if (contentType && /\bapplication\/json\b/.test(contentType)) {
-          if (error.body == null) {
-            data = null;
-          } else {
-            data = await error.json();
-          }
-        } else {
-          data = await error.text();
-        }
-
+        let data = await unwrapResponse(error);
         errorForSecondRender = new ErrorResponseImpl(
           error.status,
           error.statusText,
@@ -377,21 +575,35 @@ async function handleDocumentRequestRR(
       context.errors = sanitizeErrors(context.errors, serverMode);
     }
 
-    // Update entryContext for the second render pass
+    // Get a new entryContext for the second render pass
+    // Server UI state to send to the client.
+    // - When single fetch is enabled, this is streamed down via `serverHandoffStream`
+    // - Otherwise it's stringified into `serverHandoffString`
+    let state = {
+      loaderData: context.loaderData,
+      actionData: context.actionData,
+      errors: serializeErrors(context.errors, serverMode),
+    };
     entryContext = {
       ...entryContext,
       staticHandlerContext: context,
       serverHandoffString: createServerHandoffString({
-        url: context.location.pathname,
         basename: build.basename,
-        state: {
-          loaderData: context.loaderData,
-          actionData: context.actionData,
-          errors: serializeErrors(context.errors, serverMode),
-        },
         future: build.future,
         isSpaMode: build.isSpaMode,
+        ...(!build.future.v3_singleFetch ? { state } : null),
       }),
+      ...(build.future.v3_singleFetch
+        ? {
+            serverHandoffStream: encodeViaTurboStream(
+              state,
+              request.signal,
+              build.entry.module.streamTimeout,
+              serverMode
+            ),
+            renderMeta: {},
+          }
+        : null),
     };
 
     try {
@@ -409,8 +621,9 @@ async function handleDocumentRequestRR(
   }
 }
 
-async function handleResourceRequestRR(
+async function handleResourceRequest(
   serverMode: ServerMode,
+  build: ServerBuild,
   staticHandler: StaticHandler,
   routeId: string,
   request: Request,
@@ -425,12 +638,27 @@ async function handleResourceRequestRR(
       routeId,
       requestContext: loadContext,
     });
-    invariant(
-      !(DEFERRED_SYMBOL in response),
-      `You cannot return a \`defer()\` response from a Resource Route.  Did you ` +
-        `forget to export a default UI component from the "${routeId}" route?`
-    );
-    // callRouteLoader/callRouteAction always return responses
+
+    if (typeof response === "object" && response !== null) {
+      invariant(
+        !(DEFERRED_SYMBOL in response),
+        `You cannot return a \`defer()\` response from a Resource Route.  Did you ` +
+          `forget to export a default UI component from the "${routeId}" route?`
+      );
+    }
+
+    if (build.future.v3_singleFetch && !isResponse(response)) {
+      console.warn(
+        resourceRouteJsonWarning(
+          request.method === "GET" ? "loader" : "action",
+          routeId
+        )
+      );
+      response = json(response);
+    }
+
+    // callRouteLoader/callRouteAction always return responses (w/o single fetch).
+    // With single fetch, users should always be Responses from resource routes
     invariant(
       isResponse(response),
       "Expected a Response to be returned from queryRoute"
@@ -440,8 +668,8 @@ async function handleResourceRequestRR(
     if (isResponse(error)) {
       // Note: Not functionally required but ensures that our response headers
       // match identically to what Remix returns
-      error.headers.set("X-Remix-Catch", "yes");
-      return error;
+      let response = safelySetHeader(error, "X-Remix-Catch", "yes");
+      return response;
     }
 
     if (isRouteErrorResponse(error)) {
@@ -490,4 +718,60 @@ function returnLastResortErrorResponse(error: any, serverMode?: ServerMode) {
       "Content-Type": "text/plain",
     },
   });
+}
+
+function unwrapResponse(response: Response) {
+  let contentType = response.headers.get("Content-Type");
+  // Check between word boundaries instead of startsWith() due to the last
+  // paragraph of https://httpwg.org/specs/rfc9110.html#field.content-type
+  return contentType && /\bapplication\/json\b/.test(contentType)
+    ? response.body == null
+      ? null
+      : response.json()
+    : response.text();
+}
+
+function createRemixRedirectResponse(
+  response: Response,
+  basename: string | undefined
+) {
+  // We don't have any way to prevent a fetch request from following
+  // redirects. So we use the `X-Remix-Redirect` header to indicate the
+  // next URL, and then "follow" the redirect manually on the client.
+  let headers = new Headers(response.headers);
+  let redirectUrl = headers.get("Location")!;
+  headers.set(
+    "X-Remix-Redirect",
+    basename ? stripBasename(redirectUrl, basename) || redirectUrl : redirectUrl
+  );
+  headers.set("X-Remix-Status", String(response.status));
+  headers.delete("Location");
+  if (response.headers.get("Set-Cookie") !== null) {
+    headers.set("X-Remix-Revalidate", "yes");
+  }
+
+  return new Response(null, {
+    status: 204,
+    headers,
+  });
+}
+
+// Anytime we are setting a header on a `Response` created in the loader/action,
+// we have to so it in this manner since in an `undici` world, if the `Response`
+// came directly from a `fetch` call, the headers are immutable will throw if
+// we try to set a new header.  This is a sort of shallow clone of the `Response`
+// so we can safely set our own header.
+function safelySetHeader(
+  response: Response,
+  name: string,
+  value: string
+): Response {
+  let headers = new Headers(response.headers);
+  headers.set(name, value);
+  return new Response(response.body, {
+    status: response.status,
+    statusText: response.statusText,
+    headers,
+    duplex: response.body ? "half" : undefined,
+  } as ResponseInit & { duplex?: "half" });
 }